+++ navigator/r2-navigator-swift/EPUB/EPUBSpreadView.swift
@@ -31,9 +31,6 @@ protocol EPUBSpreadViewDelegate: AnyObject {
     /// Called when the spread view needs to present a view controller.
     func spreadView(_ spreadView: EPUBSpreadView, present viewController: UIViewController)
 
-    /// Called when the spread view receives an unknown JavaScript message.
-    func spreadView(_ spreadView: EPUBSpreadView, userContentController: WKUserContentController, didReceive message: WKScriptMessage)
-    
 }
 
 class EPUBSpreadView: UIView, Loggable, PageView {
@@ -398,11 +395,10 @@ extension EPUBSpreadView: WKScriptMessageHandler {
 
     /// Handles incoming calls from JS.
     func userContentController(_ userContentController: WKUserContentController, didReceive message: WKScriptMessage) {
-        if let handler = JSMessages[message.name]  {
-            handler(message.body)
-        } else {
-            delegate?.spreadView(self, userContentController: userContentController, didReceive: message)
+        guard let handler = JSMessages[message.name] else {
+            return
         }
+        handler(message.body)
     }
 
 }